Shopping

+++ agricultural-crop-management-backend/src/main/java/org/example/QuanLyMuaVu/module/marketplace/dto/response/MarketplaceCartItemResponse.java
@@ -8,8 +8,8 @@ public record MarketplaceCartItemResponse(
         String name,
         String imageUrl,
         BigDecimal unitPrice,
-        Integer quantity,
-        Integer maxQuantity,
+        BigDecimal quantity,
+        BigDecimal maxQuantity,
         Long farmerUserId,
         Boolean traceable) {
 }
